requirements.txt
pkg-resources==0.0.0
PyYAML==5.1
